+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added experimental `filter_data` option
 - Localized times on maintenance page
+- Improved connection pooling
 
 ## 2.4.2 (2020-04-16)